+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Add CacheProxy which caches the storage in memory
 
 ## 0.1.0 - 2016-12-07
 ### Added